Improved lsseq version numbering. Now following 'Semaintic Versioning 2.0.0' standard.
That is, M.m.p where,
  - M is Major version
  - m is minor version
  - p is patch version

M, m and p are non-padded positive integers, with no breaks in the sequence.

My OLD convention was, M.mpp where
  - M was Major version
  - m was minor version
  - pp, first p was patch version, second p was mini-change to patch.

As such, I renamed all old tarballs as best as I could, so the version numbers
of the tarballs match (minus some dots (.)) their old numbers, which means that
the version number printed by the 'lsseq --version' (that is, the one contained in
each tarball) still uses the old naming convention. (I'm not going to recreate the
tarballs!!)

Note that my renaming of the tarballs didn't leave a list of files that
adhere properly to the standard, There are gaps in patch-numbering as I used
to use the last two digits as patch and minor-patch numbers but the renaming
is good enough for me to move forward correctly in the future.

In the latest version, that is 2.1.0, lsseq correctly prints out this
exact number with appropriate dots when you use the --version option.

I'm goint to do an official "release" on github, and just learned about this
versioning standard, so decided to adopt it. Mine was a little sloppy before, but
close to the convention. From now on the convention will be followed to the letter.