Deciding to tag this new version as v3.0.0 even though technically it's backward compatible with v2.7.1 as far as the end user sees.
However this new version requires that the version of python available
be >= python3.7, where as lsseq-v2.7.1 only required >= python3.6. However
since the end-of-life support for python3.6 was the end of 2021, which
is 2 1/2 years ago, it seems fairly safe to assume that at least python3.7
is available, given that it's end-of-life has also passed in mid-2023.

If for some reason a system is stuck at py3.6, with no higher
version available, then those users don't have to upgrade to
lsseq-v3.0.0, and can stay parked on lsseq-v2.7.1. The improvements
made in v3.0.0 are only going to be required by a new release of renumseq
which is introducing a new --rename option. Most people won't notice any
difference between lsseq-v2.7.1 and lsseq-v3.0.0