Added two tests to test_lsseq and checking in the current output of lsseq.
The test confirms what I suspected, that is, when doing an "lsseq -R -P"
versus "lsseq -R -P *" in the same directory (in the test dir 'pdir'),
the first case prepends a "./" to each entry of the output, whereas the
second does not. This seems correct as far as the code is concerned, but
doesn't seem desireable as far as what lsseq should output.

I think --prependPathRel should NEVER print "./" at the head of any
line of output. It's redundant, clutters up the output, and doesn't
add any helpful information.

I think this qualifies as a bug, that is, this exact behavior has been
overlooked until now, but it seems that the desired output of --prepentPathRel
should NOT prepend a "./", like the second case when using a "*" glob pattern.

Stripping "./" from the head of a path should not break any scripts that might
use this option, as the path with or without "./" refers to the same set of files.

I'm going to fix this bug, and generate the test output again, and make a new
release once tested.