The option --prependPathRel should never have printed "./" at the head of any line of output. It's redundant, clutters up the output, and doesn't add any helpful information.
This is a design-bug fix, and with this fix 'lsseq -R -P' and 'lsseq -R -P *'
(i.e. with the globbing wildcard) NOW produce the exact same resuts in any
given directory from which lsseq is run.

Stripping "./" from the head of the relative path will not break any scripts
that might be using this option, because the path with or without "./" refers to
the exact same set of files.