First bugfix of a couple, plus a new feature (coming) for the next release.
Shake is a long defunct program so this change will almost CERTAINLY not
affect anyone in the world.

When using '--format shake' the output was including the
command 'shake' followed by ' -t <frame range> <SEQ>'.  Not sure why
I wrote it that way. I expect it was a hold over from VERY old version of
lsseq when shake was still in use (yes, lsseq is that old) and it was
convenient to output the line that way. It's not consistent however
with the other format outputs.

At any rate, it's much more correct to omit the word "shake" and
flip the order of the line such that it prints the sequence name
FOLLOWED by '-t <frame range>'. This is consistent with shake
command line usage too, as I just consulted an old manual.

So this commit achieves that change and updates the test output with
the new results.