Implemented the --split-sequence option. Tested and confirmed working, as well as all regression tests passed.
Also fixed a bug in the '--format nuke' option. The program 'nuke', say in the
read-node dialog window, when encountering a sequence with missing frames,
will append the actual sequence (showing only the frames present) in
parenthesis at the end of the line. Now lsseq also prints this extra
list of frames in parenthesis when frames are missing in the sequence.